refactor: rename Aside as ArticleInfo
